+++ src/Api/Dfe.Complete.Api/Controllers/TasksDataController.cs
@@ -160,6 +160,25 @@ namespace Dfe.Complete.Api.Controllers
         }
 
         /// <summary>
+        /// Updates the TUPE Consultation task data for conversion project.
+        /// </summary>
+        /// <param name="request">The update command.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [Authorize(Policy = "CanReadWriteUpdate")]
+        [HttpPatch]
+        [Route("TaskData/TupeConsultation")]
+        [SwaggerResponse(204, "Conversion's TUPE Consultation task updated successfully.")]
+        [SwaggerResponse(400, "Invalid request data.")]
+        [SwaggerResponse(404, "Project/User not found.")]
+        public async Task<IActionResult> UpdateTupeConsultationTaskAsync(
+            [FromBody] UpdateTupeConsultationTaskCommand request,
+            CancellationToken cancellationToken)
+        {
+            await sender.Send(request, cancellationToken);
+            return NoContent();
+        }
+
+        /// <summary>
         /// Updates the redact and send documents task Data for conversion or transfer project.
         /// </summary>
         /// <param name="request">The update command.</param>